Move stuff to generic directories
